# go faster by caching writes per thread and merging as we go.