--- a/cpp/structs.hpp
+++ b/cpp/structs.hpp
@@ -2,9 +2,6 @@
 #include <vector>
 #include <unordered_set>
 #include <mutex>
-#include <shared_mutex>
-#include <atomic>
-#include <map>
 
 using namespace std;
 // #define DBG 1