@nsch0e 2023-07-14
is there a reason why you put hashy into namespace std? I think separate sets per shape is more interesting, because by clearing the cache we accept cubes which are already in the total set and the rejection just comes later when we merge the cache...