@nsch0e 2023-07-14
please checkout my branch "subhashy". There I group sets by same shape.
Results for N = 12, Threads = 16:
cpp: ~81s
yours: ~50s
subhashy: ~26s

is subhashy also faster for you?