@bertie2 2023-07-14
agreed the multisets are they way to go but have an improvement and more comments, closing this pr and raising a more relevant one.